# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BridgeCore Platform API
  description: >
    BridgeCore is a reusable multi-tenant backend platform providing
    authentication, authorization, feature entitlements, API keys, usage
    metering, and audit logging as REST APIs for downstream products.
  version: "0.1.0"
  contact:
    name: BridgeCore Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [auth]
      summary: Register a new tenant and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_name, email, password]
              properties:
                tenant_name: { type: string }
                email: { type: string }
                password: { type: string, minLength: 8 }
                first_name: { type: string }
                last_name: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: Bad request }
        "409": { description: Email or slug already taken }
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Invalid credentials }
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Invalid or expired refresh token }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke all refresh tokens for the current user
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get the currently authenticated user
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
...
  /api/v1/features:
    get:
      tags: [entitlements]
      summary: List the full feature catalog
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/features/mine:
    get:
      tags: [entitlements]
      summary: List features enabled for the current tenant
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/features/grant:
    post:
      tags: [entitlements]
      summary: Grant or revoke a feature for a tenant (admin only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, feature_key, enabled]
              properties:
                tenant_id: { type: string }
                feature_key: { type: string }
                enabled: { type: boolean }
      responses:
        "200": { description: OK }
  /api/v1/apikeys:
    post:
      tags: [apikeys]
      summary: Generate a new API key
      security: [{ BearerAuth: [] }]
      responses:
        "201": { description: Created }
    get:
      tags: [apikeys]
      summary: List API keys for the current tenant
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/apikeys/{id}/rotate:
    post:
      tags: [apikeys]
      summary: Rotate an API key
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/v1/apikeys/{id}:
    delete:
      tags: [apikeys]
      summary: Deactivate (revoke) an API key
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /api/v1/audit:
    get:
      tags: [audit]
      summary: List audit log entries for the current tenant
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: event
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        "200": { description: OK }
  /api/v1/audit/{id}:
    get:
      tags: [audit]
      summary: Get a single audit log entry by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Not found }